Add files via upload
A pandas Series can be queried either by the index position or the index label. If you don't give an  index to the series when querying, the position and the label are effectively the same values. To query by numeric location, starting at zero, use the iloc attribute. To query by the index label,  you can use the loc attribute.